Face-Verifier-Server/requirements.txt
tensorflow==2.13.1
numpy==1.23.5
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
Pillow==9.4.0
keras==2.13.1
Flask==2.2.5
gdown==4.7.1
kserve==0.11.2